backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pandas==2.1.3
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0
python-jose==3.3.0
python-dotenv==1.0.0
websockets==12.0
google-generativeai==0.3.1
openai==1.3.7
pydantic==2.5.0
pydantic-settings==2.1.0